@@ -16,35 +16,36 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^5.2.6",
-    "@angular/common": "^5.2.6",
-    "@angular/compiler": "^5.2.6",
-    "@angular/core": "^5.2.6",
-    "@angular/forms": "^5.2.6",
-    "@angular/http": "^5.2.6",
-    "@angular/platform-browser": "^5.2.6",
-    "@angular/platform-browser-dynamic": "^5.2.6",
-    "@angular/platform-server": "^5.2.6",
-    "@angular/router": "^5.2.6",
-    "@angular/upgrade": "^5.2.6",
+    "@angular/animations": "^7.2.3",
+    "@angular/common": "^7.2.3",
+    "@angular/compiler": "^7.2.3",
+    "@angular/core": "^7.2.3",
+    "@angular/forms": "^7.2.3",
+    "@angular/http": "^7.2.3",
+    "@angular/platform-browser": "^7.2.3",
+    "@angular/platform-browser-dynamic": "^7.2.3",
+    "@angular/platform-server": "^7.2.3",
+    "@angular/router": "^7.2.3",
+    "@angular/upgrade": "^7.2.3",
     "angular2-jwt": "0.2.3",
     "chart.js": "2.5.0",
     "core-js": "2.4.1",
     "moment": "^2.24.0",
-    "ngx-bootstrap": "^2.0.2",
-    "rxjs": "^5.5.6",
+    "ngx-bootstrap": "^2.0.5",
+    "rxjs": "^6.4.0",
     "ts-helpers": "1.1.2",
-    "tslib": "^1.9.0",
-    "zone.js": "^0.8.20"
+    "tslib": "^1.9.3",
+    "zone.js": "^0.8.29",
+    "rxjs-compat": "^6.0.0-rc.0"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~0.12.0",
-    "@angular/cli": "^7.2.3",
-    "@angular/compiler-cli": "^5.2.6",
+    "@angular-devkit/build-angular": "^0.12.4",
+    "@angular/cli": "^7.3.0",
+    "@angular/compiler-cli": "^7.2.3",
     "@types/jasmine": "2.5.45",
     "@types/node": "7.0.8",
     "autotrack": "^2.4.1",
-    "codelyzer": "^4.1.0",
+    "codelyzer": "^4.5.0",
     "jasmine-core": "2.5.2",
     "jasmine-spec-reporter": "3.2.0",
     "karma": "^4.0.0",
@@ -55,7 +56,7 @@
     "karma-jasmine-html-reporter": "0.2.2",
     "protractor": "^5.4.2",
     "ts-node": "2.1.0",
-    "tslint": "^5.9.1",
-    "typescript": "^2.7.2"
+    "tslint": "^5.12.1",
+    "typescript": "^3.2.4"
   }
 }
